requirements/base.txt
Django==1.5.1
ipython==1.1.0
django-model-utils==1.4.0
logutils==0.3.3
psycopg2==2.5.1
South==0.8.1
Fabric==1.8.0
#PIL==1.1.7
Pillow==2.2.1
django-userena==1.2.4
#django-authtools=0.2.1
django-pagedown=0.0.5
django-sslify==0.2
